Refactor: Rename module to hold all the falcon api models
