+++ fig/falcon/__init__.py
@@ -1,5 +1,5 @@
 from .api import FalconAPI
-from .event import Event
+from .models import Event
 from .stream import StreamManagementThread
 
 __all__ = ['Event', 'FalconAPI', 'StreamManagementThread']
